% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloom_survey}
\alias{bloom_survey}
\title{2017 Bloomington, Indiana Community survey}
\format{A data frame with 595 rows and 227 variables:
\describe{
  \item{id}{not defined in data source}
  \item{wave}{not defined in data source}
  \item{area}{Zip Code}
  \item{q1a}{Bloomington as a place to live}
  \item{q1b}{Your neighborhood as a place to live}
  \item{q1c}{Bloomington as a place to raise children}
  \item{q1d}{Bloomington as a place to work}
  \item{q1e}{Bloomington as a place to visit}
  \item{q1f}{Bloomington as a place to retire}
  \item{q1g}{The overall quality of life in the City of Bloomington}
  \item{q2a}{Overall feeling of safety in Bloomington}
  \item{q2b}{Overall ease of getting to the places you usually visit}
  \item{q2c}{Quality of overall natural environment in Bloomington}
  \item{q2d}{Overall Òbuilt environmentÓ of Bloomington (including overall design, of buildings, parks and transportation systems)}
  \item{q2e}{Health and wellness opportunities in Bloomington}
  \item{q2f}{Overall opportunities for education and enrichment}
  \item{q2g}{Overall economic health of Bloomington}
  \item{q2h}{Sense of community}
  \item{q2i}{Overall image or reputation of Bloomington}
  \item{q3a}{Recommend living in the City of Bloomington to someone who asks}
  \item{q3b}{Remain in the City of Bloomington for the next five years}
  \item{q4a}{In your neighborhood during the day}
  \item{q4b}{In your neighborhood at night}
  \item{q4c}{In Bloomington's downtown area during the day}
  \item{q4d}{In Bloomington's downtown area at night}
  \item{q5a}{Traffic flow on major streets}
  \item{q5b}{Ease of public parking}
  \item{q5c}{Cost of public parking}
  \item{q5d}{Ease of travel by car }
  \item{q5e}{Ease of travel by public transportation }
  \item{q5f}{Ease of travel by bicycle }
  \item{q5g}{Ease of walking }
  \item{q5h}{Availability of paths and walking trails}
  \item{q5i}{Air quality}
  \item{q5j}{Cleanliness of the City}
  \item{q5k}{Overall appearance}
  \item{q5l}{Public places where people want to spend time}
  \item{q5m}{Variety of housing options}
  \item{q5n}{Availability of affordable housing}
  \item{q5o}{Fitness opportunities}
  \item{q5p}{Recreational opportunities}
  \item{q5q}{Availability of affordable food}
  \item{q5r}{Availability of affordable health care}
  \item{q5s}{Availability of affordable mental health care}
  \item{q6a}{Availability of affordable child care/preschool}
  \item{q6b}{K-12 education}
  \item{q6c}{Adult educational opportunities}
  \item{q6d}{Opportunities to attend cultural/arts/music activities}
  \item{q6e}{Opportunities to participate in religious or spiritual events and activities}
  \item{q6f}{Employment opportunities}
  \item{q6g}{Shopping opportunities downtown}
  \item{q6h}{Shopping opportunities outside downtown}
  \item{q6i}{Cost of living in Bloomington}
  \item{q6j}{Overall quality of business and service establishments in Bloomington}
  \item{q6k}{Entertainment downtown}
  \item{q6l}{Entertainment outside downtown}
  \item{q6m}{Overall development of a variety of housing options in downtown (i.e. townhomes, apartments, etc.)}
  \item{q6n}{Opportunities to participate in social events and activities}
  \item{q6o}{Opportunities to volunteer}
  \item{q6p}{Opportunities to participate in community matters}
  \item{q6q}{Openness and acceptance of the community toward people of diverse backgrounds}
  \item{q6r}{Neighborliness of residents in Bloomington}
  \item{q7a}{Made efforts to conserve water}
  \item{q7b}{Made efforts to make your home more energy efficient}
  \item{q7c}{Observed a code violation or other hazard in the City of Bloomington (weeds, scattered trash, etc.)}
  \item{q7d}{You or a household member were a victim of a crime in the City of Bloomington}
  \item{q7e}{You or a household member reported a crime to the police in the City of Bloomington}
  \item{q7f}{Stocked supplies in preparation for an emergency}
  \item{q7g}{Campaigned or advocated for an issue, cause or candidate}
  \item{q7h}{Attended an event, meeting or convention at the Bloomington/Monroe Convention Center}
  \item{q8}{Have you contacted any City of Bloomington staff (parks, police, public works, parking, etc.) in-person, by phone, email, web or uReport within the last 12 months?}
  \item{q9a}{Knowledgeable}
  \item{q9b}{Responsive}
  \item{q9c}{Courteous}
  \item{q9d}{Overall impression}
  \item{q10a}{Attended a local public meeting }
  \item{q10b}{Watched (online or on television) a local public meeting}
  \item{q11a}{Participated in religious or spiritual activities}
  \item{q11b}{Attended an event offered by the City of Bloomington}
  \item{q11c}{Used bus or other public transportation instead of driving}
  \item{q11d}{Carpooled with other adults or children instead of driving alone}
  \item{q11e}{Walked or biked instead of driving}
  \item{q11f}{Volunteered your time to some group/activity in Bloomington}
  \item{q11g}{Talked to or visited with your immediate neighbors}
  \item{q11h}{Done a favor for a neighbor}
  \item{q12a}{The City of Bloomington}
  \item{q12b}{Indiana State Government}
  \item{q12c}{The Federal Government}
  \item{q13a}{The value of services for the taxes paid to City}
  \item{q13b}{The overall direction that the City is taking}
  \item{q13c}{The job City government does at welcoming citizen involvement}
  \item{q13d}{Overall confidence in City government}
  \item{q13e}{Generally acting in the best interest of the community}
  \item{q13f}{Being honest/transparent}
  \item{q13g}{Treating all residents fairly}
  \item{q13h}{Listening to public concern}
  \item{q14a}{Switchyard Park completion}
  \item{q14b}{Development/job creation in the Trades District/certified tech park}
  \item{q14c}{Expansion of the Bloomington/Monroe County Convention Center}
  \item{q14d}{Redevelopment of current hospital site on W. 2nd St.}
  \item{q14e}{City-wide high speed fiber for internet access}
  \item{q15}{To what degree, if at all, do you think homelessness is a challenge in the City of Bloomington?}
  \item{q16}{Supportive housing combines housing with services such as job training and substance abuse programs. To what extent would you support or oppose the City assisting supportive housing for residents who are experiencing homelessness? }
  \item{q17a}{City website (Bloomington.in.gov)}
  \item{q17b}{Local media outlets (newspapers, radio, local television stations)}
  \item{q17c}{The local government television channel, CATs}
  \item{q17d}{City Council meetings and other public meetings}
  \item{q17e}{Talking with City officials}
  \item{q17f}{City communications via social media (i.e. Facebook, Twitter etc.)}
  \item{q17g}{Word-of-mouth}
  \item{q17h}{Email distribution lists}
  \item{q18a}{Police services}
  \item{q18b}{Fire services}
  \item{q18c}{Ambulance or emergency medical service}
  \item{q18d}{Crime prevention}
  \item{q18e}{Fire prevention and education}
  \item{q18f}{Traffic enforcement}
  \item{q18g}{Street repair}
  \item{q18h}{Street cleaning}
  \item{q18i}{Street lighting}
  \item{q18j}{Snow removal}
  \item{q18k}{Sidewalk maintenance}
  \item{q18l}{Downtown parking}
  \item{q18m}{Neighborhood/residential parking}
  \item{q18n}{Parking garages}
  \item{q18o}{Traffic signal timing}
  \item{q18p}{Bus or transit services}
  \item{q18q}{Garbage collection}
  \item{q18r}{Recycling}
  \item{q18s}{Yard waste pick-up}
  \item{q18t}{Storm drainage}
  \item{q18u}{Drinking water}
  \item{q18v}{Sewer services}
  \item{q18w}{Land use, planning and zoning}
  \item{q18x}{Code enforcement (weeds, trash, etc.)}
  \item{q18y}{Animal care and control}
  \item{q18z}{Economic development}
  \item{q18aa}{Sustainability efforts}
  \item{q18bb}{Health services}
  \item{q18cc}{Public information services}
  \item{q18dd}{Preservation of natural areas (i.e. forests)}
  \item{q18ee}{Open space (green space)}
  \item{q18ff}{FarmersÕ market}
  \item{q18gg}{Taste of Bloomington}
  \item{q18hh}{Lotus Festival}
  \item{q18ii}{Fourth Street Arts Fair}
  \item{q18jj}{Overall customer service by City Employees}
  \item{q18kk}{City website}
  \item{q18ll}{Affordable/workforce housing availability}
  \item{q18mm}{Resident interaction with the Bloomington Police Department}
  \item{q18nn}{Public art provided by the City}
  \item{q18oo}{Small Business support}
  \item{q18pp}{Parks}
  \item{q18qq}{Trails}
  \item{q18aimp}{Police services}
  \item{q18bimp}{Fire services}
  \item{q18cimp}{Ambulance or emergency medical service}
  \item{q18dimp}{Crime prevention}
  \item{q18eimp}{Fire prevention and education}
  \item{q18fimp}{Traffic enforcement}
  \item{q18gimp}{Street repair}
  \item{q18himp}{Street cleaning}
  \item{q18iimp}{Street lighting}
  \item{q18jimp}{Snow removal}
  \item{q18kimp}{Sidewalk maintenance}
  \item{q18limp}{Downtown parking}
  \item{q18mimp}{Neighborhood/residential parking}
  \item{q18nimp}{Parking garages}
  \item{q18oimp}{Traffic signal timing}
  \item{q18pimp}{Bus or transit services}
  \item{q18qimp}{Garbage collection}
  \item{q18rimp}{Recycling}
  \item{q18simp}{Yard waste pick-up}
  \item{q18timp}{Storm drainage}
  \item{q18uimp}{Drinking water}
  \item{q18vimp}{Sewer services}
  \item{q18wimp}{Land use, planning and zoning}
  \item{q18ximp}{Code enforcement (weeds, trash, etc.)}
  \item{q18yimp}{Animal care and control}
  \item{q18zimp}{Economic development}
  \item{q18aaimp}{Sustainability efforts}
  \item{q18bbimp}{Health services}
  \item{q18ccimp}{Public information services}
  \item{q18ddimp}{Preservation of natural areas (i.e. forests)}
  \item{q18eeimp}{Open space (green space)}
  \item{q18ffimp}{FarmersÕ market}
  \item{q18ggimp}{Taste of Bloomington}
  \item{q18hhimp}{Lotus Festival}
  \item{q18iiimp}{Fourth Street Arts Fair}
  \item{q18jjimp}{Overall customer service by City Employees}
  \item{q18kkimp}{City website}
  \item{q18llimp}{Affordable/workforce housing availability}
  \item{q18mmimp}{Resident interaction with the Bloomington Police Department}
  \item{q18nnimp}{Public art provided by the City}
  \item{q18ooimp}{Small Business support}
  \item{q18ppimp}{Parks}
  \item{q18qqimp}{Trails}
  \item{d1a}{Recycle at home}
  \item{d1b}{Purchase goods or services from a business located in Bloomington}
  \item{d1c}{Read or watch local news (via television, paper, computer, etc.)}
  \item{d1d}{Vote in local elections}
  \item{d2}{What is your employment status?}
  \item{d3}{To what extent do you agree or disagree that your household is currently experiencing financial security?}
  \item{d4}{Do you work inside the boundaries of the City of Bloomington?}
  \item{d5}{How many years have you lived in the City of Bloomington? }
  \item{d6}{Which best describes the building you live in?}
  \item{d7}{Do you rent or own your home?}
  \item{d8}{About how much is your monthly housing cost for the place you live (including rent, mortgage payment, property tax, property insurance and homeownersÕ association (HOA) fees)?}
  \item{d9}{Do any children 17 or under live in your household?}
  \item{d10}{Are you or any other members of your household aged 65 or older?}
  \item{d11}{How much do you anticipate your householdÕs total income before taxes will be for the current year? (Please include in your total income money from all sources for all persons living in your household.)}
  \item{d12}{Are you Spanish, Hispanic or Latino?}
  \item{d13a}{American Indian or Alaskan Native}
  \item{d13b}{Asian, Asian Indian or Pacific Islander}
  \item{d13c}{Black or African American}
  \item{d13d}{White}
  \item{d13e}{Other}
  \item{d14}{In which category is your age?}
  \item{d15}{With what gender do you identify most closely?}
  \item{d16}{Do you consider a cell phone or land line your primary telephone number?}
  \item{d17}{Are you a student?}
  \item{mode}{not defined in data source}
  \item{wt}{not defined in data source}
  \item{q19}{Q19 verbatim coded by survey collectors}
  \item{q20}{Q20 verbatim coded by survey collectors}
  \item{q19verb}{What is the one thing you like least about the City of Bloomington?}
  \item{q20verb}{What is the one thing you like most about the City of Bloomington?}
}}
\source{
\url{https://catalog.data.gov/dataset/community-survey}
}
\usage{
bloom_survey
}
\description{
A dataset containing survey responses from Bloomington, Indiana's
community survey in 2017. Two open-ended questions are utilized.
}
\keyword{datasets}
