% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{create_phrasefreq}
\alias{create_phrasefreq}
\title{Create phrase frequency}
\usage{
create_phrasefreq(data)
}
\arguments{
\item{data}{Data frame with the unnested phrases}
}
\description{
Creates data frame with count of occurences of each phrase in each column.
This function will be called by analysis functions, not intended to be
used directly.
}
