% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sentiment.R
\name{plot_response_sentiment}
\alias{plot_response_sentiment}
\title{Creates exploratory plots of the assignment of sentiments to words}
\usage{
plot_response_sentiment(data)
}
\arguments{
\item{data}{Data frame assigned sentiments (typically from response_sentiment())}
}
\description{
Creates and prints ggplots showing how the sentiment assignment works.
One or more plots can be displayed depending on the user preference.
}
\details{
The first plot, "sentiment_histogram", shows a histogram for each sentiment
and question. It shows the distribution of the sentiment by the frequency
of occurences
}
