% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_sentiment.R
\name{response_sentiment}
\alias{response_sentiment}
\title{Assign sentiments to unigrams using nrc, afinn and bing dictionaries}
\usage{
response_sentiment(tidy_data)
}
\arguments{
\item{tidy_data}{Data frame with the unnested phrases (usually from tidy_verbatim())}
}
\description{
Creates data frame with one row per survey, selected question, and summarized 
sentiment. Three sentiment dictionaries are used; the first time this function runs
you will need to download the dictionary (there will be a prompt in the console).
}
\details{
Afinn and Bing dictionaries will each produce a single score which represents the 
overall polarity of sentiment. These dictionaries score each applicable word for 
its negativity or positivity; Afinn scores from -5 to 5, while Bing just labels as 
negative/positive. The Afinn index is the sum of all individual word scores and the 
Bing index represents the count of positive words minus count of negative words.

The NRC dictionary assigns words to one of several sentiments, such as anger or joy.
The summarized values will represent the count of words assigned to each sentiment
on the response and question level.
}
