% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_verbatim.R
\name{tidy_verbatim}
\alias{tidy_verbatim}
\title{Create a tidy data frame with phrases from each column}
\usage{
tidy_verbatim(data, id = id, cols = selected_cols, n = 1)
}
\arguments{
\item{data}{Data frame with the survey data}

\item{id}{Unique survey id. By default, can be omitted if column is named 'id'}

\item{cols}{List of column(s) to be unnested. By default, can be omitted if 
using vector named 'selected_cols'}

\item{n}{Length of phrases, in words. By default is 1 (unigrams)}
}
\value{
A data frame with one row per phrase and column, with survey id.
}
\description{
This function creates a long table with one row per phrase and column.
It allows the user to set the n-gram parameter (how many words per phrase)
and then utilizes the tidytext library to unnest each verbatim response.
Any NA responses are dropped.
}
\details{
It's recommended that you select your columns into a vector outside
of the function call, because you will probably want to re-use that
list with other functions in the package.
}
\examples{
data(bloom_survey)
selected_cols <- c("q19verb", "q20verb")
surveys <- tidy_verbatim(bloom_survey, n = 3)
}
